moved some statics to class member

+++ Yuyuyui.PrivateServer.DataModel/Gifts/Gift.cs
@@ -8,7 +8,7 @@ namespace Yuyuyui.PrivateServer.DataModel
         public long Id { get; set; }
         public long ContentId { get; set; }
         public string ContentType { get; set; } = "";
-        public long Quantity { get; set; }
+        public int Quantity { get; set; }
         public string Title { get; set; } = "";
     }
 }
